+++ py_css_styleguide/__init__.py
@@ -1,2 +1,2 @@
 """CSS Manifest driven styleguide for your project"""
-__version__ = '0.3.0-pre.6'
+__version__ = '0.3.0'