+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 Version 0.3.0 - Unreleased
 --------------------------
 
+* Internally use ``collection.OrderedDict`` instead of simple dictionnary in parser and serializer, close #4;
 * Changed ``list`` and ``string`` structures so they can be empty, close #3;
 * Fixed Sass helper function ``to-string`` for empty list, close #2;